backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
sqlalchemy
python-multipart
cython
pydantic
python-dotenv
